Fix typo

+++ src/qibolab/platforms/abstract.py
@@ -218,7 +218,7 @@ class AbstractPlatform(ABC):
                 elif "rotation_angle" in par:
                     self.qubits[qubit].rotation_angle = float(value)
                 elif "length" in par:  # assume only drive length
-                    self.native_single_qubit_gates[qubit]["RX"]["duration"] = float(value)
+                    self.native_single_qubit_gates[qubit]["RX"]["duration"] = int(value)
                 else:
                     raise_error(ValueError, "Unknown parameter.")
